requirements.txt
langgraph
langchain
langchain-google-genai
mcp
requests
python-dotenv